backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
# psycopg2-binary==2.9.9  # Commented out for Windows compatibility
alembic==1.13.1

# AI/ML Libraries
langchain==0.1.6
google-generativeai==0.8.5
langchain-google-genai==3.0.0
chromadb==0.4.22
sentence-transformers==2.3.1
transformers==4.36.2
torch==2.1.2
whisper==1.1.10

# Document Processing
pymupdf==1.23.21
openpyxl==3.1.2
python-docx==1.1.0
camelot-py==0.11.0
tabula-py==2.8.2
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.3.2
prophet==1.1.5

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Voice Processing
SpeechRecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1
edge-tts==6.1.9
soundfile==0.12.1

# Web Scraping & HTTP
requests==2.31.0
httpx==0.26.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Background Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Storage
boto3==1.34.0
minio==7.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.8.0

# Security
cryptography==41.0.7
python-jose==3.3.0
passlib==1.7.4

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Additional Dependencies
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
click==8.1.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

